requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bleach==3.3.0
boto3==1.17.106
botocore==1.20.106
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-rest-auth==2.1.7
Django==3.2.5
django-allauth==0.44.0
django-autoslug==1.9.8
django-baton==2.2.1
django-bleach==0.6.1
django-bootstrap4==3.0.1
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.1
django-multiselectfield==0.1.12
django-mysql==3.12.0
django-storages==1.11.1
django-taggit==1.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
idna==2.10
jmespath==0.10.0
oauthlib==3.1.1
packaging==21.0
Pillow==8.3.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
webencodings==0.5.1
